+++ phoenix-scala/phoenix/app/phoenix/models/cord/CordBase.scala
@@ -11,9 +11,9 @@ trait CordBase[A <: FoxModel[A]] extends FoxModel[A] { self: A ⇒
   def accountId: Int
   def currency: Currency
 
-  def subTotal: Int
-  def shippingTotal: Int
-  def adjustmentsTotal: Int
-  def taxesTotal: Int
-  def grandTotal: Int
+  def subTotal: Long
+  def shippingTotal: Long
+  def adjustmentsTotal: Long
+  def taxesTotal: Long
+  def grandTotal: Long
 }
