+++ phoenix-scala/phoenix/app/phoenix/models/cord/Order.scala
@@ -19,16 +19,17 @@ import utils.Money.Currency
 import utils.db.ExPostgresDriver.api._
 import utils.db._
 
+
 case class Order(id: Int = 0,
                  scope: LTree,
                  referenceNumber: String = "",
                  accountId: Int,
                  currency: Currency = Currency.USD,
-                 subTotal: Int = 0,
-                 shippingTotal: Int = 0,
-                 adjustmentsTotal: Int = 0,
-                 taxesTotal: Int = 0,
-                 grandTotal: Int = 0,
+                 subTotal: Long = 0,
+                 shippingTotal: Long = 0,
+                 adjustmentsTotal: Long = 0,
+                 taxesTotal: Long = 0,
+                 grandTotal: Long = 0,
                  // Order-specific
                  contextId: Int,
                  state: Order.State = Order.RemorseHold,
@@ -94,11 +95,11 @@ class Orders(tag: Tag) extends FoxTable[Order](tag, "orders") {
   def referenceNumber  = column[String]("reference_number")
   def accountId        = column[Int]("account_id")
   def currency         = column[Currency]("currency")
-  def subTotal         = column[Int]("sub_total")
-  def shippingTotal    = column[Int]("shipping_total")
-  def adjustmentsTotal = column[Int]("adjustments_total")
-  def taxesTotal       = column[Int]("taxes_total")
-  def grandTotal       = column[Int]("grand_total")
+  def subTotal         = column[Long]("sub_total")
+  def shippingTotal    = column[Long]("shipping_total")
+  def adjustmentsTotal = column[Long]("adjustments_total")
+  def taxesTotal       = column[Long]("taxes_total")
+  def grandTotal       = column[Long]("grand_total")
   def contextId        = column[Int]("context_id")
   def state            = column[Order.State]("state")
   def placedAt         = column[Instant]("placed_at")
