+++ phoenix-scala/phoenix/app/phoenix/models/cord/lineitems/CartLineItemAdjustment.scala
@@ -7,6 +7,7 @@ import phoenix.utils.ADT
 import shapeless._
 import slick.ast.BaseTypedType
 import slick.jdbc.JdbcType
+import utils.Money._
 import utils.db.ExPostgresDriver.api._
 import utils.db._
 
@@ -14,7 +15,7 @@ final case class CartLineItemAdjustment(id: Int = 0,
                                         cordRef: String,
                                         promotionShadowId: Int,
                                         adjustmentType: CartLineItemAdjustment.AdjustmentType,
-                                        subtract: Int,
+                                        subtract: Long,
                                         lineItemRefNum: Option[String] = None,
                                         createdAt: Instant = Instant.now)
     extends FoxModel[CartLineItemAdjustment]
@@ -26,7 +27,7 @@ class CartLineItemAdjustments(tag: Tag)
   def cordRef           = column[String]("cord_ref")
   def promotionShadowId = column[Int]("promotion_shadow_id")
   def adjustmentType    = column[CartLineItemAdjustment.AdjustmentType]("adjustment_type")
-  def subtract          = column[Int]("subtract")
+  def subtract          = column[Long]("subtract")
   def lineItemRefNum    = column[Option[String]]("line_item_ref_num")
   def createdAt         = column[Instant]("created_at")
