+++ phoenix-scala/phoenix/app/phoenix/services/LogActivity.scala
@@ -56,6 +56,7 @@ import phoenix.services.activity.UserTailored._
 import phoenix.utils.aliases._
 import responses.ObjectResponses.ObjectContextResponse
 import utils.db._
+import utils.Money._
 
 case class LogActivity(implicit ac: AC) {
 
@@ -243,11 +244,11 @@ case class LogActivity(implicit ac: AC) {
                                        GiftCardResponse.build(gc),
                                        StoreCreditResponse.build(sc)))
 
-  def gcFundsAuthorized(user: User, cart: Cart, gcCodes: Seq[String], amount: Int)(
+  def gcFundsAuthorized(user: User, cart: Cart, gcCodes: Seq[String], amount: Long)(
       implicit ec: EC): DbResultT[Activity] =
     Activities.log(GiftCardAuthorizedFunds(buildUser(user), cart, gcCodes, amount))
 
-  def gcFundsCaptured(user: User, order: Order, gcCodes: Seq[String], amount: Int)(
+  def gcFundsCaptured(user: User, order: Order, gcCodes: Seq[String], amount: Long)(
       implicit ec: EC): DbResultT[Activity] =
     Activities.log(GiftCardCapturedFunds(buildUser(user), order, gcCodes, amount))
 
@@ -268,11 +269,11 @@ case class LogActivity(implicit ac: AC) {
                                        GiftCardResponse.build(gc),
                                        StoreCreditResponse.build(sc)))
 
-  def scFundsAuthorized(user: User, cart: Cart, scIds: Seq[Int], amount: Int)(
+  def scFundsAuthorized(user: User, cart: Cart, scIds: Seq[Int], amount: Long)(
       implicit ec: EC): DbResultT[Activity] =
     Activities.log(StoreCreditAuthorizedFunds(buildUser(user), cart, scIds, amount))
 
-  def scFundsCaptured(user: User, order: Order, scIds: Seq[Int], amount: Int)(
+  def scFundsCaptured(user: User, order: Order, scIds: Seq[Int], amount: Long)(
       implicit ec: EC): DbResultT[Activity] =
     Activities.log(StoreCreditCapturedFunds(buildUser(user), order, scIds, amount))
 
@@ -371,7 +372,7 @@ case class LogActivity(implicit ac: AC) {
                                          CreditCardsResponse.build(cc, region),
                                          buildOriginator(originator)))
 
-  def orderPaymentMethodAddedGc(originator: User, cart: CartResponse, gc: GiftCard, amount: Int)(
+  def orderPaymentMethodAddedGc(originator: User, cart: CartResponse, gc: GiftCard, amount: Long)(
       implicit ec: EC): DbResultT[Activity] =
     Activities.log(
         CartPaymentMethodAddedGiftCard(cart,
@@ -382,8 +383,8 @@ case class LogActivity(implicit ac: AC) {
   def orderPaymentMethodUpdatedGc(originator: User,
                                   cart: CartResponse,
                                   gc: GiftCard,
-                                  oldAmount: Option[Int],
-                                  amount: Int)(implicit ec: EC): DbResultT[Activity] = {
+                                  oldAmount: Option[Long],
+                                  amount: Long)(implicit ec: EC): DbResultT[Activity] = {
     val activity = CartPaymentMethodUpdatedGiftCard(cart,
                                                     GiftCardResponse.build(gc),
                                                     oldAmount,
@@ -393,7 +394,7 @@ case class LogActivity(implicit ac: AC) {
     Activities.log(activity)
   }
 
-  def orderPaymentMethodAddedSc(originator: User, cart: CartResponse, amount: Int)(
+  def orderPaymentMethodAddedSc(originator: User, cart: CartResponse, amount: Long)(
       implicit ec: EC): DbResultT[Activity] =
     Activities.log(CartPaymentMethodAddedStoreCredit(cart, amount, buildOriginator(originator)))
